Fixed #36248 -- Limited delete-time related-model updates to batch size
During deletion, when there is a relation pointing to the deleted model with an on_delete mode that requires an update operation (such as SET_NULL), do one update statement for each batch of deletes that get_del_batches has calculated, instead of combining them into one filter. This should prevent us hitting the parameter limit in DBMSs that have one (eg SQLite, MSSQL).

Add a test along the lines of the DeletionTests.test_large_delete, counting the number of queries done to infer that we must have split the update into multiple queries. This isn't as explicit as I would like but it seems to be the best we can do given it's not currently possible to inspect the number of parameters used in the query log, and that we can't rely on SQLite exploding on >999 parameters any more (since newer versions have higher limits, and the actual configured value of the limit isn't immediately visible to us).